Crate prisma_models

Re-exports

pub extern crate datamodel;

pub use datamodel::dml;

Modules

arithmetic

Note: Only number operations are implemented at the moment.

pk

prelude

Structs

CompositeField

CompositeIndexField

CompositeSelection

CompositeType

ConversionFailure

FieldSelection

A selection of fields from a model.

Fields

Index

InlineRelation

InternalDataModel

InternalDataModelBuilder

InternalEnum

InternalEnumValue

ManyRecords

Model

ModelProjection

Projection of a Model. A projection is a (sub)set of fields of a model. There can only ever be fields of one model contained in a particular ModelProjection

OrderByRelevance

OrderByScalar

OrderByScalarAggregation

OrderByToManyAggregation

Record

Relation

A relation between two models. Can be either using a RelationTable or model a direct link between two RelationFields.

RelationField

RelationTable

ScalarField

SelectionResult

Represents a set of results.

SingleRecord

Enums

DateType

DomainError

Field

FieldWeak

IndexField

IndexType

NullsOrder

OrderBy

OrderByHop

Describes a hop over to a relation or composite for an orderBy statement.

PrismaValue

RelationLinkManifestation

RelationSide

SelectedField

A selected field. Can be contained on a model or composite type.

SortAggregation

SortOrder

TypeIdentifier

Traits

PrismaValueExtensions

Functions

decode_bytes

encode_bytes

stringify_date

Stringify a date to the following format 1999-05-01T00:00:00.000Z

Type Definitions

CompositeFieldRef

CompositeFieldWeak

CompositeTypeRef

CompositeTypeWeakRef

InternalDataModelRef

InternalDataModelWeakRef

InternalEnumRef

ModelRef

ModelWeakRef

PrismaListValue

PrismaValueResult

RelationFieldRef

RelationFieldWeak

RelationRef

RelationWeakRef

Result

ScalarFieldRef

ScalarFieldWeak