Struct prisma_models::ast::SchemaAst

pub struct SchemaAst {
    pub tops: Vec<Top, Global>,
}

AST representation of a prisma schema.

This module is used internally to represent an AST. The AST’s nodes can be used during validation of a schema, especially when implementing custom attributes.

The AST is not validated, also fields and attributes are not resolved. Every node is annotated with its location in the text representation. Basically, the AST is an object oriented representation of the datamodel’s text. Schema = Datamodel + Generators + Datasources

Fields
tops: Vec<Top, Global>

All models, enums, composite types, datasources, generators and type aliases.

Implementations

impl SchemaAst

pub fn find_at_position(&self, position: usize) -> SchemaPosition<'_>

Find the AST node at the given position (byte offset).

pub fn find_top_at_position(&self, position: usize) -> Option<TopId>

Do a binary search for the Top at the given byte offset.

impl SchemaAst

pub fn iter_tops(&self) -> impl Iterator<Item = (TopId, &Top)>

Iterate over all the top-level items in the schema.

pub fn sources(&self) -> impl Iterator<Item = &SourceConfig>

Iterate over all the datasource blocks in the schema.

pub fn generators(&self) -> impl Iterator<Item = &GeneratorConfig>

Iterate over all the generator blocks in the schema.