Module prisma_models::dml

Modules

prisma_value

Structs

CompositeField

Represents a composite field.

CompositeType

CompositeTypeField

Datamodel

Entities in the datamodel can be flagged as is_commented_out. This lets the renderer know that introspection encountered unsupported names or features and these are supposed to be rendered as comments. Since the parser will not set these flags when reading a schema string, only introspection and the lowering of the datamodel to the ast care about these flags. The FieldType: Unsupported behaves in the same way. Both of these are never converted into the internal datamodel.

DefaultValue

Represents a default specified on a field.

Enum

Represents an enum in the datamodel.

EnumValue

Represents a value of an enum

IndexDefinition

Represents an index defined via @@index, @unique or @@unique.

IndexField

A field in an index that optionally defines a sort order and length limit.

Model

Represents a model in a prisma schema.

NativeTypeInstance

represents an instance of a native type declared in the Prisma schema

PrimaryKeyDefinition

Represents a primary key defined via @@id or @id.

PrimaryKeyField

A field in a Primary Key that optionally defines a sort order and length limit.

RelationField

Represents a relation field in a model.

RelationInfo

Holds information about a relation field.

ScalarField

Represents a scalar field in a model.

UniqueCriteria

A unique criteria is a set of fields through which a record can be uniquely identified.

ValueGenerator

Enums

CompositeTypeFieldType

DefaultKind

Represents a default specified on a field.

Field

Represents a Field in a Model.

FieldArity

Arity of a Field in a Model.

FieldType

Datamodel field type.

IndexAlgorithm

IndexType

OperatorClass

PrismaValue

ReferentialAction

Describes what happens when related nodes are deleted.

ScalarType

Prisma’s builtin scalar types.

SortOrder

ValueGeneratorFn

Traits

Ignorable

WithDatabaseName

WithName

Functions

find_model_by_db_name

Find the model mapping to the passed in database name.