Struct prisma_models::dml::Model

pub struct Model {
    pub name: String,
    pub fields: Vec<Field, Global>,
    pub documentation: Option<String>,
    pub database_name: Option<String>,
    pub indices: Vec<IndexDefinition, Global>,
    pub primary_key: Option<PrimaryKeyDefinition>,
    pub is_generated: bool,
    pub is_commented_out: bool,
    pub is_ignored: bool,
}

Represents a model in a prisma schema.

Fields
name: String

Name of the model.

fields: Vec<Field, Global>

Fields of the model.

documentation: Option<String>

Comments associated with this model.

database_name: Option<String>

The database internal name of this model.

indices: Vec<IndexDefinition, Global>

Describes Composite Indexes

primary_key: Option<PrimaryKeyDefinition>

Describes the Primary Keys

is_generated: bool

Indicates if this model is generated.

is_commented_out: bool

Indicates if this model has to be commented out.

is_ignored: bool

Indicates if this model has to be ignored by the Client.

Implementations

impl Model

pub fn new(name: String, database_name: Option<String>) -> Model

Creates a new model with the given name.

pub fn add_field(&mut self, field: Field)

Adds a field to this model.

pub fn fields(&self) -> Iter<'_, Field>

Gets an iterator over all fields.

pub fn fields_mut(&mut self) -> impl Iterator<Item = &mut Field>

Gets a mutable iterator over all fields.

pub fn scalar_fields(&self) -> impl Iterator<Item = &ScalarField>

Gets an iterator over all scalar fields.

pub fn relation_fields(&self) -> impl Iterator<Item = &RelationField>

Gets an iterator over all relation fields.

pub fn scalar_fields_mut(&mut self) -> impl Iterator<Item = &mut ScalarField>

Gets a mutable iterator over all scalar fields.

pub fn relation_fields_mut(
&mut self
) -> impl Iterator<Item = &mut RelationField>

Gets a mutable iterator over all relation fields.

pub fn find_field(&self, name: &str) -> Option<&Field>

Finds a field by name.

pub fn find_field_mut(&mut self, name: &str) -> &mut Field

Finds a field by name and returns a mutable reference.

pub fn find_scalar_field(&self, name: &str) -> Option<&ScalarField>

Finds a scalar field by name.

pub fn find_relation_field(&self, name: &str) -> Option<&RelationField>

Finds a scalar field by name.

pub fn find_scalar_field_db_name(&self, db_name: &str) -> Option<&ScalarField>

Finds a field by database name.

pub fn has_field(&self, name: &str) -> bool

pub fn find_scalar_field_mut(&mut self, name: &str) -> &mut ScalarField

Finds a field by name and returns a mutable reference.

pub fn find_relation_field_mut(&mut self, name: &str) -> &mut RelationField

Finds a relation field by name and returns a mutable reference.

pub fn first_unique_criterion(&self) -> Vec<&ScalarField, Global>

This should match the logic in prisma_models::Model::primary_identifier.

pub fn strict_unique_criterias(&self) -> Vec<UniqueCriteria<'_>, Global>

optional unique fields are NOT considered a unique criteria used for: A Model must have at least one STRICT unique criteria.

pub fn strict_unique_criterias_disregarding_unsupported(
&self
) -> Vec<UniqueCriteria<'_>, Global>

optional unique fields are NOT considered a unique criteria used for: A Model must have at least one STRICT unique criteria. Ignores unsupported, used for introspection to decide when to ignore

pub fn loose_unique_criterias(&self) -> Vec<UniqueCriteria<'_>, Global>

optional unique fields are considered a unique criteria used for: A relation must reference one LOOSE unique criteria. (optional fields are okay in this case)

pub fn field_is_indexed(&self, field_name: &str) -> bool

pub fn has_single_id_field(&self) -> bool

Determines whether there is a singular primary key

pub fn add_index(&mut self, index: IndexDefinition)

pub fn has_created_at_and_updated_at(&self) -> bool

pub fn field_is_unique(&self, name: &str) -> bool

pub fn field_is_unique_and_defined_on_field(&self, name: &str) -> bool

pub fn field_is_primary(&self, field_name: &str) -> bool

pub fn field_is_primary_and_defined_on_field(&self, field_name: &str) -> bool

pub fn field_is_auto_generated_int_id(&self, name: &str) -> bool

pub fn id_field_names(&self) -> Vec<String, Global>