Module prisma_models::prelude

Re-exports

pub use crate::sql_ext::*;

Modules

arithmetic

Note: Only number operations are implemented at the moment.

dml

Structs

ConversionFailure

DatamodelConverter

Fields

Index

IndexTemplate

InlineRelation

InternalDataModel

InternalDataModelTemplate

InternalEnum

InternalEnumValue

ManyRecords

Model

ModelProjection

Projection of a Model. A projection is a (sub)set of fields of a model. There can only ever be fields of one model contained in a particular ModelProjection

ModelTemplate

OrderByAggregation

OrderByRelevance

OrderByScalar

Record

RecordProjection

Represents a (sub)set of fields to value pairs from a single record.

Relation

A relation between two models. Can be either using a RelationTable or model a direct link between two RelationFields.

RelationField

RelationFieldTemplate

RelationTable

RelationTemplate

ScalarField

ScalarFieldTemplate

SingleRecord

TempRelationHolder

Enums

DateType

DomainError

Field

FieldBehaviour

FieldTemplate

FieldWeak

IndexType

OrderBy

PrismaValue

RelationLinkManifestation

RelationSide

ScalarListStrategy

SortAggregation

SortOrder

TempManifestationHolder

TypeIdentifier

Functions

decode_bytes

encode_bytes

stringify_date

Stringify a date to the following format 1999-05-01T00:00:00.000Z

Type Definitions

InternalDataModelRef

InternalDataModelWeakRef

InternalEnumRef

ModelRef

ModelWeakRef

PrismaListValue

PrismaValueResult

RelationFieldRef

A short-hand for Arc<RelationField>

RelationFieldWeak

A short-hand for Weak<RelationField>

RelationRef

RelationWeakRef

ScalarFieldRef

ScalarFieldWeak