Struct prisma_models::prelude::FieldSelection

pub struct FieldSelection { /* fields omitted */ }

A selection of fields from a model.

Implementations

impl FieldSelection

pub fn new(selections: Vec<SelectedField>) -> Self

pub fn into_inner(self) -> Vec<SelectedField>

pub fn is_superset_of(&self, other: &Self) -> bool

Returns true if self contains (at least) all fields specified in other. false otherwise. Recurses into composite selections and ensures that composite selections are supersets as well.

pub fn selections(&self) -> impl Iterator<Item = &SelectedField> + '_

pub fn prisma_names(&self) -> impl Iterator<Item = String> + '_

Returns all Prisma (e.g. schema model field) names of contained fields. Does not recurse into composite selections and only iterates top level fields.

pub fn db_names(&self) -> impl Iterator<Item = String> + '_

Returns all database (e.g. column or document field) names of contained fields. Does not recurse into composite selections and only iterates level fields.

pub fn contains(&self, name: &str) -> bool

Checked if a field of prisma name name is present in this FieldSelection.

pub fn get(&self, name: &str) -> Option<&SelectedField>

pub fn as_fields(&self) -> Vec<Field>

pub fn as_scalar_fields(&self) -> Option<Vec<ScalarFieldRef>>

Checks if self only contains scalar field selections and if so, returns them all in a list. If any other selection is contained, returns None.

pub fn assimilate(&self, values: SelectionResult) -> Result<SelectionResult>

Inserts this FieldSelections selections into the given SelectionResult. Assumes caller knows that the exchange can be done, but still errors if lengths mismatch. Additionally performs a type coercion based on the source and destination field types. Resistance is futile.

pub fn matches(&self, result: &SelectionResult) -> bool

Checks if a given SelectionResult belongs to this FieldSelection.

pub fn union(selections: Vec<Self>) -> Self

Merges all given FieldSelection a set union of all. Each selection is contained exactly once, with the first occurrence of the first field in order from left to right is retained.

/!\ Important assumption: All selections are on the same model.

pub fn merge(self, other: FieldSelection) -> FieldSelection

Consumes both FieldSelections to create a new one that contains a union of both. Each selection is contained exactly once, with the first occurrence of the first field in order from left (self) to right (other) is retained. Assumes that both selections reason over the same model.