Struct prisma_models::prelude::Fields

pub struct Fields { /* private fields */ }

Implementations

impl Fields

pub fn id(&self) -> Option<PrimaryKey>

pub fn compound_id(&self) -> Option<PrimaryKey>

pub fn updated_at(&self) -> impl Iterator<Item = ScalarFieldRef>

pub fn scalar(&self) -> Vec<ScalarFieldRef>

pub fn relation(&self) -> Vec<RelationFieldRef>

pub fn composite(&self) -> Vec<CompositeFieldRef>

pub fn non_relational(&self) -> Vec<Field>

pub fn find_many_from_scalar(
&self,
names: &BTreeSet<String>
) -> Vec<ScalarFieldRef>

pub fn find_from_all(&self, prisma_name: &str) -> Result<Field>

pub fn find_from_non_virtual_by_db_name(&self, db_name: &str) -> Result<Field>

Non-virtual: Fields actually existing on the database level, this (currently) excludes relations, which are purely virtual on a model.

pub fn find_from_scalar(&self, name: &str) -> Result<ScalarFieldRef>

pub fn find_from_relation_fields(&self, name: &str) -> Result<RelationFieldRef>

pub fn filter_all<P>(&self, predicate: P) -> Vec<Field> where
P: Fn(&&Field) -> bool,