Struct prisma_models::prelude::Model

pub struct Model {
    pub id: ModelId,
    pub internal_data_model: InternalDataModelWeakRef,
    pub supports_create_operation: bool,
    /* private fields */
}

Fields
id: ModelIdinternal_data_model: InternalDataModelWeakRefsupports_create_operation: bool
Implementations

impl Model

pub fn name(&self) -> &str

pub fn schema_name(&self) -> Option<String>

Returns the schema name for the model which is the contents of the @@schema(“…”) attribute

pub fn primary_identifier(&self) -> FieldSelection

Returns the set of fields to be used as the primary identifier for a record of that model. The identifier is nothing but an internal convention to have an anchor point for querying, or in other words, the identifier is not to be mistaken for a stable, external identifier, but has to be understood as implementation detail that is used to reason over a fixed set of fields.

pub fn fields(&self) -> &Fields

pub fn supports_create_operation(&self) -> bool

pub fn indexes(&self) -> &[Index]

pub fn unique_indexes(&self) -> Vec<&Index>

pub fn db_name(&self) -> &str

The name of the model in the database For a sql database this will be the Table name for this model

pub fn db_name_opt(&self) -> Option<&str>

pub fn internal_data_model(&self) -> InternalDataModelRef