Struct prisma_models::prelude::RecordProjection

pub struct RecordProjection {
    pub pairs: Vec<(ScalarFieldRef, PrismaValue)>,
}

Represents a (sub)set of fields to value pairs from a single record.

Fields
pairs: Vec<(ScalarFieldRef, PrismaValue)>
Implementations

impl RecordProjection

pub fn new(pairs: Vec<(ScalarFieldRef, PrismaValue)>) -> Self

pub fn add(&mut self, pair: (ScalarFieldRef, PrismaValue))

pub fn model(&self) -> Option<ModelRef>

pub fn fields(&self) -> impl Iterator<Item = ScalarFieldRef> + '_

pub fn values(&self) -> impl Iterator<Item = PrismaValue> + '_

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn misses_autogen_value(&self) -> bool

pub fn add_autogen_value<V>(&mut self, value: V) -> bool where
V: Into<PrismaValue>,

pub fn split_into(
self,
projections: &[ModelProjection]
) -> Vec<RecordProjection>

Consumes this projection and splits it into a set of RecordProjections based on the passed ModelProjections. Assumes that the transformation can be done.

pub fn ensure_type_coherence(self) -> Self

Ensures that the contained values match the type identifier.