Struct prisma_models::Relation

pub struct Relation {
    pub name: String,
    pub manifestation: RelationLinkManifestation,
    pub internal_data_model: InternalDataModelWeakRef,
    // some fields omitted
}

A relation between two models. Can be either using a RelationTable or model a direct link between two RelationFields.

Fields
name: Stringmanifestation: RelationLinkManifestationinternal_data_model: InternalDataModelWeakRef
Implementations

impl Relation

pub const MODEL_A_DEFAULT_COLUMN: &'static str

pub const MODEL_B_DEFAULT_COLUMN: &'static str

pub const TABLE_ALIAS: &'static str

pub fn is_inline_relation(&self) -> bool

Returns true only if the Relation is just a link between two RelationFields.

pub fn is_relation_table(&self) -> bool

Returns true if the Relation is a table linking two models.

pub fn is_self_relation(&self) -> bool

A model that relates to itself. For example a Person that is a parent can relate to people that are children.

pub fn model_a(&self) -> ModelRef

A pointer to the first Model in the Relation.

pub fn model_b(&self) -> ModelRef

A pointer to the second Model in the Relation.

pub fn field_a(&self) -> RelationFieldRef

A pointer to the RelationField in the first Model in the Relation.

pub fn field_b(&self) -> RelationFieldRef

A pointer to the RelationField in the second Model in the Relation.

pub fn is_many_to_many(&self) -> bool

Practically deprecated with Prisma 2.

pub fn is_one_to_one(&self) -> bool

pub fn is_one_to_many(&self) -> bool

pub fn internal_data_model(&self) -> InternalDataModelRef

pub fn on_delete(&self) -> ReferentialAction

Retrieves the onDelete policy for this relation.

pub fn on_update(&self) -> ReferentialAction

Retrieves the onUpdate policy for this relation.