Struct prisma_models::RelationField

pub struct RelationField {
    pub arity: FieldArity,
    pub relation_name: String,
    pub relation_side: RelationSide,
    pub relation: OnceCell<RelationWeakRef>,
    pub on_delete_default: ReferentialAction,
    pub on_update_default: ReferentialAction,
    /* private fields */
}

Fields
arity: FieldArityrelation_name: Stringrelation_side: RelationSiderelation: OnceCell<RelationWeakRef>on_delete_default: ReferentialActionon_update_default: ReferentialAction
Implementations

impl RelationField

pub fn name(&self) -> &str

pub fn is_list(&self) -> bool

pub fn is_required(&self) -> bool

pub fn linking_fields(&self) -> FieldSelection

Returns the FieldSelection used for this relation fields model.

What is the field selection of a relation field?

The set of fields required by the relation (on the model of the relation field) to be able to link the related records.

In case of a many-to-many relation field, we can make the assumption that the primary identifier of the enclosing model is the set of linking fields, as this is how Prisma many-to-many works and we only support implicit join tables (i.e. m:n) in the Prisma style.

pub fn is_optional(&self) -> bool

pub fn model(&self) -> ModelRef

pub fn scalar_fields(&self) -> Vec<ScalarFieldRef>

pub fn relation(&self) -> RelationRef

pub fn is_inlined_on_enclosing_model(&self) -> bool

Alias for more clarity (in most cases, doesn’t add more clarity for self-relations);

pub fn relation_is_inlined_in_parent(&self) -> bool

Inlined in self / model of self

pub fn relation_is_inlined_in_child(&self) -> bool

pub fn related_model(&self) -> ModelRef

pub fn related_field(&self) -> Arc<RelationField>

pub fn is_relation_with_name_and_side(
&self,
relation_name: &str,
side: RelationSide
) -> bool

pub fn type_identifiers_with_arities(&self) -> Vec<(TypeIdentifier, FieldArity)>

pub fn referenced_fields(&self) -> Vec<ScalarFieldRef>

pub fn left_scalars(&self) -> Vec<ScalarFieldRef>

pub fn right_scalars(&self) -> Vec<ScalarFieldRef>

pub fn db_names(&self) -> impl Iterator<Item = String>

pub fn on_delete(&self) -> Option<&ReferentialAction>

pub fn on_update(&self) -> Option<&ReferentialAction>

pub fn relation_info(&self) -> &RelationInfo