Struct prisma_models::Zipper

pub struct Zipper<I> {
    pub id: I,
    pub dm: InternalDataModelRef,
}

Fields
id: Idm: InternalDataModelRef
Implementations

impl Zipper<CompositeTypeId>

pub fn name(&self) -> &str

pub fn fields(&self) -> impl Iterator<Item = Field> + '_

pub fn find_field(&self, prisma_name: &str) -> Option<Field>

pub fn find_field_by_db_name(&self, db_name: &str) -> Option<Field>

impl Zipper<CompositeFieldId>

pub fn typ(&self) -> CompositeType

pub fn is_list(&self) -> bool

pub fn is_required(&self) -> bool

pub fn is_optional(&self) -> bool

pub fn name(&self) -> &str

pub fn db_name(&self) -> &str

pub fn container(&self) -> ParentContainer

impl Zipper<RelationFieldId>

pub fn name(&self) -> &str

pub fn arity(&self) -> FieldArity

pub fn is_list(&self) -> bool

pub fn is_required(&self) -> bool

pub fn linking_fields(&self) -> FieldSelection

Returns the FieldSelection used for this relation fields model.

What is the field selection of a relation field?

The set of fields required by the relation (on the model of the relation field) to be able to link the related records.

In case of a many-to-many relation field, we can make the assumption that the primary identifier of the enclosing model is the set of linking fields, as this is how Prisma many-to-many works and we only support implicit join tables (i.e. m:n) in the Prisma style.

pub fn is_optional(&self) -> bool

pub fn model(&self) -> ModelRef

pub fn scalar_fields(&self) -> Vec<ScalarFieldRef>

pub fn relation(&self) -> Relation

pub fn is_inlined_on_enclosing_model(&self) -> bool

Alias for more clarity (in most cases, doesn’t add more clarity for self-relations);

pub fn relation_is_inlined_in_parent(&self) -> bool

Inlined in self / model of self

pub fn relation_is_inlined_in_child(&self) -> bool

pub fn related_model(&self) -> ModelRef

pub fn related_field(&self) -> RelationField

pub fn type_identifiers_with_arities(&self) -> Vec<(TypeIdentifier, FieldArity)>

pub fn referenced_fields(&self) -> Vec<ScalarFieldRef>

pub fn left_scalars(&self) -> Vec<ScalarFieldRef>

pub fn db_names(&self) -> impl Iterator<Item = String>

impl Zipper<ScalarFieldId>

pub fn is_id(&self) -> bool

pub fn is_list(&self) -> bool

pub fn is_required(&self) -> bool

pub fn unique(&self) -> bool

pub fn db_name(&self) -> &str

pub fn type_identifier_with_arity(&self) -> (TypeIdentifier, FieldArity)

pub fn is_read_only(&self) -> bool

pub fn is_numeric(&self) -> bool

pub fn container(&self) -> ParentContainer

pub fn name(&self) -> &str

pub fn type_identifier(&self) -> TypeIdentifier

pub fn arity(&self) -> FieldArity

pub fn internal_enum(&self) -> Option<InternalEnum>

pub fn default_value(&self) -> Option<DefaultKind>

pub fn is_updated_at(&self) -> bool

pub fn is_auto_generated_int_id(&self) -> bool

pub fn native_type(&self) -> Option<NativeTypeInstance>

pub fn is_autoincrement(&self) -> bool

impl Zipper<EnumId>

pub fn name(&self) -> &str

impl Zipper<ModelId>

pub fn name(&self) -> &str

pub fn schema_name(&self) -> Option<String>

Returns the schema name for the model which is the contents of the @@schema(“…”) attribute

pub fn primary_identifier(&self) -> FieldSelection

Returns the set of fields to be used as the primary identifier for a record of that model. The identifier is nothing but an internal convention to have an anchor point for querying, or in other words, the identifier is not to be mistaken for a stable, external identifier, but has to be understood as implementation detail that is used to reason over a fixed set of fields.

pub fn fields(&self) -> Fields

pub fn supports_create_operation(&self) -> bool

pub fn db_name(&self) -> &str

The name of the model in the database For a sql database this will be the Table name for this model

pub fn db_name_opt(&self) -> Option<&str>

pub fn internal_data_model(&self) -> InternalDataModelRef

pub fn unique_indexes(&self) -> impl Iterator<Item = IndexWalker<'_>>

impl Zipper<RelationId>

pub fn name(&self) -> String

pub fn is_inline_relation(&self) -> bool

Returns true only if the Relation is just a link between two RelationFields.

pub fn is_relation_table(&self) -> bool

Returns true if the Relation is a table linking two models.

pub fn is_self_relation(&self) -> bool

A model that relates to itself. For example a Person that is a parent can relate to people that are children.

pub fn is_many_to_many(&self) -> bool

Practically deprecated with Prisma 2.

pub fn is_one_to_one(&self) -> bool

pub fn is_one_to_many(&self) -> bool

pub fn on_delete(&self) -> ReferentialAction

Retrieves the onDelete policy for this relation.

pub fn on_update(&self) -> ReferentialAction

Retrieves the onUpdate policy for this relation.

impl<I: Copy> Zipper<I>

pub fn walker(&self) -> Walker<'_, I>