Type Definition prisma_models::Relation

pub type Relation = Zipper<RelationId>;

Implementations

impl Relation

pub fn name(&self) -> String

pub fn is_inline_relation(&self) -> bool

Returns true only if the Relation is just a link between two RelationFields.

pub fn is_relation_table(&self) -> bool

Returns true if the Relation is a table linking two models.

pub fn is_self_relation(&self) -> bool

A model that relates to itself. For example a Person that is a parent can relate to people that are children.

pub fn model_a(&self) -> ModelRef

A pointer to the first Model in the Relation.

pub fn model_b(&self) -> ModelRef

A pointer to the second Model in the Relation.

pub fn field_a(&self) -> RelationFieldRef

A pointer to the RelationField in the first Model in the Relation.

pub fn field_b(&self) -> RelationFieldRef

A pointer to the RelationField in the second Model in the Relation.

pub fn is_many_to_many(&self) -> bool

Practically deprecated with Prisma 2.

pub fn is_one_to_one(&self) -> bool

pub fn is_one_to_many(&self) -> bool

pub fn on_delete(&self) -> ReferentialAction

Retrieves the onDelete policy for this relation.

pub fn on_update(&self) -> ReferentialAction

Retrieves the onUpdate policy for this relation.