Enum prisma_models::walkers::IndexFieldWalker

pub enum IndexFieldWalker<'db> {
    Scalar(Walker<'db, ScalarFieldId>),
    Composite(Walker<'db, (CompositeTypeId, FieldId)>),
}

A field in an index definition. It can point to a scalar field in the current model, or through embedding a field in a composite type.

Variants

Scalar(Walker<'db, ScalarFieldId>)

A field on a model.

Composite(Walker<'db, (CompositeTypeId, FieldId)>)

The path to a field in a composite type.

Implementations

impl<'db> IndexFieldWalker<'db>

pub fn is_optional(self) -> bool

Is the field optional / nullable?

pub fn is_list(self) -> bool

Is the field a list?

pub fn is_unsupported(self) -> bool

Is the type of the field Unsupported("...")?

pub fn field_id(self) -> FieldId

The ID of the field node in the AST.

pub fn name(self) -> &'db str

The name of the field.

pub fn database_name(self) -> &'db str

The final database name of the field. See crate docs for explanations on database names.

pub fn scalar_field_type(self) -> ScalarFieldType

The type of the field.

pub fn as_scalar_field(self) -> Option<Walker<'db, ScalarFieldId>>

Convert the walker to a scalar field, if the underlying field is in a model.

pub fn as_composite_field(
self
) -> Option<Walker<'db, (CompositeTypeId, FieldId)>>

Convert the walker to a composite field, if the underlying field is in a composite type.

pub fn is_scalar_field(self) -> bool

True if the index field is a scalar field.

pub fn is_composite_field(self) -> bool

True if the index field is a composite field.

pub fn is_single_pk(self) -> bool

Does the field define a primary key by its own.

pub fn raw_native_type(
self
) -> Option<(&'db str, &'db str, &'db [String], Span)>

(attribute scope, native type name, arguments, span)

For example: @db.Text would translate to (“db”, “Text”, &[], )

pub fn ast_field(self) -> &'db Field

The field node in the AST.