Struct prisma_models::walkers::CompleteInlineRelationWalker

pub struct CompleteInlineRelationWalker<'db> { /* private fields */ }

Represents a relation that has fields and references defined in one of the relation fields. Includes 1:1 and 1:n relations that are defined from both sides.

Implementations

impl<'db> CompleteInlineRelationWalker<'db>

pub fn referencing_model(self) -> Walker<'db, ModelId>

The model that defines the relation fields and actions.

pub fn referenced_model(self) -> Walker<'db, ModelId>

The implicit relation side.

pub fn referencing_field(self) -> Walker<'db, RelationFieldId>

pub fn referenced_field(self) -> Walker<'db, RelationFieldId>

pub fn referenced_fields(self) -> impl ExactSizeIterator + 'db

The scalar fields defining the relation on the referenced model.

pub fn referencing_fields(self) -> impl ExactSizeIterator + 'db

The scalar fields on the defining the relation on the referencing model.

pub fn on_update(self) -> ReferentialAction

Gives the onUpdate referential action of the relation. If not defined explicitly, returns the default value.

pub fn referential_arity(self) -> FieldArity

Prisma allows setting the relation field as optional, even if one of the underlying scalar fields is required. For the purpose of referential actions, we count the relation field required if any of the underlying fields is required.