Crate prisma_models

Re-exports

pub use dml;

pub use psl;

Modules

arithmetic

Note: Only number operations are implemented at the moment.

ast

The AST data structure. It aims to faithfully represent the syntax of a Prisma Schema, with source span information.

pk

prelude

walkers

Convenient access to a datamodel as understood by ParserDatabase.

Structs

CompositeSelection

ConversionFailure

FieldSelection

A selection of fields from a model.

Fields

InternalDataModel

ManyRecords

ModelProjection

Projection of a Model. A projection is a (sub)set of fields of a model. There can only ever be fields of one model contained in a particular ModelProjection

OrderByRelevance

OrderByScalar

OrderByScalarAggregation

OrderByToManyAggregation

Record

SelectionResult

Represents a set of results.

SingleRecord

Zipper

Enums

CompositeFieldId

DateType

DomainError

Field

NullsOrder

OrderBy

OrderByHop

Describes a hop over to a relation or composite for an orderBy statement.

PrismaValue

ScalarFieldId

SelectedField

A selected field. Can be contained on a model or composite type.

SortAggregation

SortOrder

TypeIdentifier

Functions

convert

decode_bytes

encode_bytes

parse_datetime

Parses an RFC 3339 and ISO 8601 date and time string such as 1996-12-19T16:39:57-08:00, then returns a new DateTime with a parsed FixedOffset.

stringify_datetime

Stringify a date to the following format 1999-05-01T00:00:00.000Z

Type Definitions

CompositeField

CompositeFieldRef

CompositeType

InternalDataModelRef

InternalEnum

InternalEnumValue

Model

ModelRef

PrismaListValue

PrismaValueResult

Relation

RelationField

RelationFieldRef

Result

ScalarField

ScalarFieldRef