Module prisma_models::prelude

Modules

arithmetic

Note: Only number operations are implemented at the moment.

Structs

CompositeSelection

ConversionFailure

FieldSelection

A selection of fields from a model.

Fields

InternalDataModel

ManyRecords

ModelProjection

Projection of a Model. A projection is a (sub)set of fields of a model. There can only ever be fields of one model contained in a particular ModelProjection

OrderByRelevance

OrderByScalar

OrderByScalarAggregation

OrderByToManyAggregation

Record

SelectionResult

Represents a set of results.

SingleRecord

Enums

CompositeFieldId

DateType

DomainError

Field

NullsOrder

OrderBy

OrderByHop

Describes a hop over to a relation or composite for an orderBy statement.

ParentContainer

PrismaValue

ScalarFieldId

SelectedField

A selected field. Can be contained on a model or composite type.

SortAggregation

SortOrder

TypeIdentifier

Functions

decode_bytes

dml_default_kind

encode_bytes

parse_datetime

Parses an RFC 3339 and ISO 8601 date and time string such as 1996-12-19T16:39:57-08:00, then returns a new DateTime with a parsed FixedOffset.

stringify_datetime

Stringify a date to the following format 1999-05-01T00:00:00.000Z

Type Definitions

CompositeField

CompositeFieldRef

Model

PrismaListValue

PrismaValueResult

Relation

RelationField

RelationFieldRef

ScalarField

ScalarFieldRef